+++ internal/directstream/stream.go
@@ -395,7 +395,8 @@ func loadContentType(path string, reader ...io.ReadSeekCloser) string {
 	case ".mp4":
 		return "video/mp4"
 	case ".mkv":
-		return "video/x-matroska"
+		//return "video/x-matroska"
+		return "video/webm"
 	case ".webm", ".m4v":
 		return "video/webm"
 	case ".avi":